# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification for Rachador — expense splitting app
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Grupo {
  id: number;
  nome: string;
  codigoConvite: string;
  criadoEm: string;
}

export interface Participante {
  id: number;
  nome: string;
  /** @nullable */
  chavePix?: string | null;
  grupoId: number;
  criadoEm: string;
}

export interface GrupoDetalhado {
  id: number;
  nome: string;
  codigoConvite: string;
  criadoEm: string;
  participantes: Participante[];
}

export interface GrupoInput {
  /** @minLength 1 */
  nome: string;
}

export interface ParticipanteInput {
  /** @minLength 1 */
  nome: string;
  chavePix?: string;
}

export interface ParticipanteUpdate {
  /** @minLength 1 */
  nome?: string;
  /** @nullable */
  chavePix?: string | null;
}

export interface Despesa {
  id: number;
  descricao: string;
  valor: number;
  pagoPorId: number;
  grupoId: number;
  criadoEm: string;
}

export interface DivisaoComParticipante {
  id: number;
  despesaId: number;
  participanteId: number;
  valorDevido: number;
  quitado: boolean;
  participante: Participante;
}

export interface DespesaDetalhada {
  id: number;
  descricao: string;
  valor: number;
  pagoPorId: number;
  grupoId: number;
  criadoEm: string;
  pagoPor: Participante;
  divisoes: DivisaoComParticipante[];
}

export interface DivisaoItem {
  participanteId: number;
  valorDevido: number;
}

export interface DespesaInput {
  /** @minLength 1 */
  descricao: string;
  /** @minimum 0.01 */
  valor: number;
  pagoPorId: number;
  /** @minItems 1 */
  participantes: DivisaoItem[];
}

export interface Divisao {
  id: number;
  despesaId: number;
  participanteId: number;
  valorDevido: number;
  quitado: boolean;
}

export interface SaldoParticipante {
  participante: Participante;
  totalPago: number;
  totalDevido: number;
  saldoLiquido: number;
}

export interface Transferencia {
  de: Participante;
  para: Participante;
  valor: number;
}

export interface SaldoGrupo {
  totalGasto: number;
  participantes: SaldoParticipante[];
  transferencias: Transferencia[];
}

export interface ResumoGrupo {
  totalDespesas: number;
  totalGasto: number;
  totalParticipantes: number;
  despesasPorParticipante: number;
  ultimasDespesas: DespesaDetalhada[];
}