# lib/api-zod/src/generated/types/despesaDetalhada.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification for Rachador — expense splitting app
 * OpenAPI spec version: 0.1.0
 */
import type { DivisaoComParticipante } from './divisaoComParticipante';
import type { Participante } from './participante';

export interface DespesaDetalhada {
  id: number;
  descricao: string;
  valor: number;
  pagoPorId: number;
  grupoId: number;
  criadoEm: Date;
  pagoPor: Participante;
  divisoes: DivisaoComParticipante[];
}