# lib/api-zod/src/generated/types/grupoDetalhado.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification for Rachador — expense splitting app
 * OpenAPI spec version: 0.1.0
 */
import type { Participante } from './participante';

export interface GrupoDetalhado {
  id: number;
  nome: string;
  codigoConvite: string;
  criadoEm: Date;
  participantes: Participante[];
}