# lib/api-zod/src/generated/types/participanteInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification for Rachador — expense splitting app
 * OpenAPI spec version: 0.1.0
 */

export interface ParticipanteInput {
  /** @minLength 1 */
  nome: string;
  chavePix?: string;
}