# lib/api-zod/src/generated/types/saldoGrupo.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification for Rachador — expense splitting app
 * OpenAPI spec version: 0.1.0
 */
import type { SaldoParticipante } from './saldoParticipante';
import type { Transferencia } from './transferencia';

export interface SaldoGrupo {
  totalGasto: number;
  participantes: SaldoParticipante[];
  transferencias: Transferencia[];
}