# lib/api-zod/src/generated/types/transferencia.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification for Rachador — expense splitting app
 * OpenAPI spec version: 0.1.0
 */
import type { Participante } from './participante';

export interface Transferencia {
  de: Participante;
  para: Participante;
  valor: number;
}